Moved requests service to helper functions
