+++ extensions/amp-story-interactive/0.1/amp-story-interactive-abstract.js
@@ -32,6 +32,7 @@ import {dict} from '#core/types/object';
 import {emojiConfetti} from './interactive-confetti';
 import {toArray} from '#core/types/array';
 import {isExperimentOn} from '#experiments/';
+import {executeRequest} from 'extensions/amp-story/1.0/story-requests';
 
 /** @const {string} */
 const TAG = 'amp-story-interactive';
@@ -174,9 +175,6 @@ export class AmpStoryInteractive extends AMP.BaseElement {
     /** @public {../../../src/service/localizationService} */
     this.localizationService = null;
 
-    /** @protected {?../../amp-story/1.0/amp-story-request-service.AmpStoryRequestService} */
-    this.requestService_ = null;
-
     /** @protected {?../../amp-story/1.0/amp-story-store-service.AmpStoryStoreService} */
     this.storeService_ = null;
 
@@ -255,9 +253,6 @@ export class AmpStoryInteractive extends AMP.BaseElement {
       Services.storyStoreServiceForOrNull(this.win).then((service) => {
         this.storeService_ = service;
         this.updateStoryStoreState_(null);
-      }),
-      Services.storyRequestServiceForOrNull(this.win).then((service) => {
-        this.requestService_ = service;
       }),
       Services.storyAnalyticsServiceForOrNull(this.win).then((service) => {
         this.analyticsService_ = service;
@@ -714,9 +709,9 @@ export class AmpStoryInteractive extends AMP.BaseElement {
         url = appendPathToUrl(this.urlService_.parse(url), ':vote');
       }
       url = addParamsToUrl(url, requestParams);
-      return this.requestService_
-        .executeRequest(url, requestOptions)
-        .catch((err) => dev().error(TAG, err));
+      return executeRequest(this.element, url, requestOptions).catch((err) =>
+        dev().error(TAG, err)
+      );
     });
   }
