+++ extensions/amp-story-shopping/0.1/amp-story-shopping-config.js
@@ -2,6 +2,8 @@ import {Layout_Enum} from '#core/dom/layout';
 
 import {Services} from '#service';
 
+import {getElementConfig} from 'extensions/amp-story/1.0/story-requests';
+
 import {
   Action,
   ShoppingConfigDataDef,
@@ -18,8 +20,6 @@ export class AmpStoryShoppingConfig extends AMP.BaseElement {
   /** @param {!AmpElement} element */
   constructor(element) {
     super(element);
-    /** @private @const {?../../amp-story/1.0/amp-story-request-service.AmpStoryRequestService} */
-    this.requestService_ = null;
     /** @private @const {?../../amp-story/1.0/amp-story-store-service.AmpStoryStoreService} */
     this.storeService_ = null;
   }
@@ -43,14 +43,11 @@ export class AmpStoryShoppingConfig extends AMP.BaseElement {
     super.buildCallback();
     return Promise.all([
       Services.storyStoreServiceForOrNull(this.win),
-      Services.storyRequestServiceForOrNull(this.win),
-    ])
-      .then(([storeService, requestService]) => {
-        this.storeService_ = storeService;
-        this.requestService_ = requestService;
-        return this.requestService_.loadConfig(this.element);
-      })
-      .then((storyConfig) => this.addShoppingDataFromConfig_(storyConfig));
+      getElementConfig(this.element),
+    ]).then(([storeService, storyConfig]) => {
+      this.storeService_ = storeService;
+      this.addShoppingDataFromConfig_(storyConfig);
+    });
   }
 
   /** @override */
